requirements.txt
aiohttp==3.8.4
openai==0.27.4
openai_whisper==20230314
pvporcupine==2.1.4
pvrecorder==1.1.1
PyAudio==0.2.13
python-dotenv==1.0.0
requests==2.28.2
whisper==1.1.10
